requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.4
django-cors-headers==3.5.0
django-environ==0.4.5
djangorestframework==3.11.0
gunicorn==20.0.4
keyboard==0.13.4
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0.1